agent-tray/requirements.txt
pystray==0.19.4
requests==2.31.0
psutil==5.9.8
Pillow==10.0.1
